My code
the code needs to connect to Database type MySQL all the information about the server and user name and password will be in the function connect_database this function you can edit the information regrading to database . for fetching the data there is another function called get_info(int a) this function it is pointer type with one variable integer this variable for select the day of the month. and it return an array with pointer. it called in the main program and it shows the data in the Text-boxes and i used an integer variable in the main program instead of the day and it increased each time when the button clicked_event invoke .